refactor: S3 관련 코드 util이 아닌 service 클래스로 변경 (#50)
